Split native terminal and browser panes
