+++ Sources/QuillCodeApp/QuillCodeBrowserPaneView.swift
@@ -1,82 +1,6 @@
 import SwiftUI
 import QuillCodeCore
 
-struct QuillCodeTerminalPaneView: View {
-    var terminal: TerminalSurface
-    @Binding var draft: String
-    var onRun: () -> Void
-    var onStop: () -> Void
-    var onClear: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            header
-            entries
-            commandLine
-        }
-        .padding(14)
-        .frame(height: 220)
-        .background(QuillCodePalette.panel)
-    }
-
-    private var header: some View {
-        HStack(spacing: 10) {
-            Image(systemName: "terminal")
-                .foregroundStyle(QuillCodePalette.blue)
-            Text("Terminal")
-                .font(.headline)
-            Text(terminal.cwdLabel)
-                .font(.caption.monospaced())
-                .foregroundStyle(QuillCodePalette.muted)
-                .lineLimit(1)
-            Spacer()
-            Button("Clear", action: onClear)
-                .controlSize(.small)
-                .disabled(!terminal.canClear)
-            if terminal.isRunning {
-                ProgressView()
-                    .controlSize(.small)
-                Button("Stop", action: onStop)
-                    .buttonStyle(.borderedProminent)
-                    .controlSize(.small)
-                    .tint(QuillCodePalette.red)
-            }
-        }
-    }
-
-    private var entries: some View {
-        ScrollView {
-            LazyVStack(alignment: .leading, spacing: 8) {
-                if terminal.entries.isEmpty {
-                    Text(terminal.emptyTitle)
-                        .font(.callout)
-                        .foregroundStyle(QuillCodePalette.muted)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .padding(.vertical, 12)
-                } else {
-                    ForEach(terminal.entries) { entry in
-                        QuillCodeTerminalEntryView(entry: entry)
-                    }
-                }
-            }
-        }
-    }
-
-    private var commandLine: some View {
-        HStack(spacing: 8) {
-            Text("$")
-                .font(.body.monospaced())
-                .foregroundStyle(QuillCodePalette.muted)
-            TextField("Run command", text: $draft)
-                .textFieldStyle(.roundedBorder)
-                .onSubmit(onRun)
-                .disabled(terminal.isRunning)
-            Button("Run", action: onRun)
-                .disabled(draft.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty || terminal.isRunning)
-        }
-    }
-}
-
 struct QuillCodeBrowserPaneView: View {
     var browser: BrowserSurface
     @Binding var addressDraft: String
@@ -330,70 +254,5 @@ struct QuillCodeBrowserPaneView: View {
         case .staticHTMLSnapshot:
             return QuillCodePalette.green
         }
-    }
-}
-
-private struct QuillCodeTerminalEntryView: View {
-    var entry: TerminalCommandSurface
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
-            HStack {
-                HStack(spacing: 8) {
-                    Text("$ \(entry.command)")
-                        .font(.caption.monospaced().weight(.semibold))
-                        .foregroundStyle(QuillCodePalette.text)
-                    if let executionContext = entry.executionContext {
-                        QuillCodeExecutionContextChip(context: executionContext)
-                    }
-                }
-                Spacer()
-                Text("\(entry.statusLabel) · \(entry.exitCodeLabel)")
-                    .font(.caption.monospacedDigit())
-                    .foregroundStyle(statusColor)
-            }
-            if !entry.stdout.isEmpty {
-                Text(entry.stdout)
-                    .font(.caption.monospaced())
-                    .textSelection(.enabled)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            if !entry.stderr.isEmpty {
-                Text(entry.stderr)
-                    .font(.caption.monospaced())
-                    .foregroundStyle(QuillCodePalette.red)
-                    .textSelection(.enabled)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-        }
-        .padding(10)
-        .background(QuillCodePalette.background.opacity(0.7))
-        .overlay(alignment: .leading) {
-            if let executionContext = entry.executionContext {
-                QuillCodeExecutionRail(context: executionContext)
-            }
-        }
-        .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-        .accessibilityLabel(accessibilityLabel)
-    }
-
-    private var statusColor: Color {
-        if entry.isSuccess {
-            return QuillCodePalette.green
-        }
-        if entry.isRunning {
-            return QuillCodePalette.blue
-        }
-        if entry.isStopped {
-            return QuillCodePalette.muted
-        }
-        return QuillCodePalette.red
-    }
-
-    private var accessibilityLabel: String {
-        let context = entry.executionContext.map {
-            ", \($0.label) \($0.detail)"
-        } ?? ""
-        return "\(entry.command), \(entry.statusLabel), \(entry.exitCodeLabel)\(context)"
     }
 }
